// src/interfaces/iTransferRecord.ts
export interface ITransferRecord {
    id: number | null;
    transferKey: string | null;
    playerUuid: string | null;
    activeFlag: string | null;
    createdWhen: string | null;
    updatedWhen: string | null;
    transferYear: number | null;
    transferDate: string | null;
    vcTargetDate: string | null;
    playerSlug: string | null;
    playerFirstName: string | null;
    playerLastName: string | null;
    playerPrefix: string | null;
    playerSuffix: string | null;
    playerHeightInches: number | null;
    playerWeightLbs: number | null;
    playerPosition: string | null;
    playerHsGradYear: number | null;
    playerHometown: string | null;
    playerHomeStateRegion: string | null;
    playerRedshirtFlag: string | null;
    playerExpectedRedshirtFlag: string | null;
    playerJucoFlag: string | null;
    playerEligibilityYear: string | null;
    fromSchoolUuid: string | null;
    fromSchoolSlug: string | null;
    fromSchoolName: string | null;
    fromSchoolNickname: string | null;
    fromSchoolCompetitiveLevel: string | null;
    fromConferenceUuid: string | null;
    fromConferenceSlug: string | null;
    fromConferenceName: string | null;
    toSchoolUuid: string | null;
    toSchoolSlug: string | null;
    toSchoolName: string | null;
    toSchoolNickname: string | null;
    toSchoolCompetitiveLevel: string | null;
    toConferenceUuid: string | null;
    toConferenceSlug: string | null;
    toConferenceName: string | null;
    playerRating: number | null;
    playerRanking: number | null;
    source: string | null;
    notes: string | null;
    highSchoolSlug: string | null;
    highSchoolName: string | null;
    highSchoolNickname: string | null;
    stats: {
        ppg: number | null;
        apg: number | null;
        rpg: number | null;
        bpg: number | null;
        spg: number | null;
        fgPct: number | null;
        ftPct: number | null;
        threePct: number | null;
        ram: number | null;
        cram: number | null;
    } | null;
}